cck-mobilya-backend-main/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-jose>=3.3
passlib>=1.7
email-validator>=2.1
boto3>=1.34
pytest>=7.4,<9
httpx>=0.25,<1
python-multipart>=0.0.9
